Revert "remove snipsync (#613)"
This reverts commit 34dcc91287435b955c803a5ab6cc947c44a36b60.
