+++ docs/snippets/conftest.py
@@ -1,14 +1,6 @@
 import os
-import pytest
-
 from tests.utils import patch_home_dir, autouse_test_storage, preserve_environ, duckdb_pipeline_location, wipe_pipeline
 
-@pytest.fixture(autouse=True)
-def set_working_dir():
-    # always set working dir to main website folder
-    abspath = os.path.abspath(__file__)
-    dname = os.path.dirname(abspath)
-    os.chdir(dname)
 
 def pytest_configure(config):
     # push sentry to ci
