+++ docs/snippets/getting_started/invoices/invoice_20230831-p1.pdf
