+++ docs/snippets/getting_started/json_file.json
