+++ docs/website/snipsync.config.yaml
@@ -1,0 +1,12 @@
+origins:
+  - files:
+    - ../snippets/**/*.py
+
+targets:
+  - docs
+
+features:
+  enable_source_link: false
+  enable_code_block: true
+  allowed_target_extensions: [.md, .mdx]
+  enable_code_dedenting: false
