+++ pytest.ini
@@ -5,5 +5,4 @@ addopts= -v --showlocals --durations 10
 xfail_strict= true
 log_cli= 1
 log_cli_level= INFO
-python_files = test_*.py *_test.py *snippets.py *snippet.pytest
-python_functions = *_test test_* *_snippet
+python_files = test_*.py *_test.py
